Add new components pack for Favorites
