+++ src/components/favorites-empty/favorites-empty.tsx
@@ -6,6 +6,7 @@ const FavoritesEmpty:React.FunctionComponent = () => {
       <div className="page__favorites-container container">
         <section className="favorites favorites--empty">
           <h1 className="visually-hidden">Favorites (empty)</h1>
+          
           <div className="favorites__status-wrapper">
             <b className="favorites__status">Nothing yet saved.</b>
             <p className="favorites__status-description">Save properties to narrow down search or plan yor future trips.</p>
